Implementado atualicao parcial de Restaurantes com metodo PATCH e uso de reflections

+++ src/main/java/com/amauri/algafood/infrastructure/repository/CidadeRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.amauri.algafood.infrastructure.repository;
 
 import com.amauri.algafood.domain.model.Cidade;
 import com.amauri.algafood.domain.repository.CidadeRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,8 +39,11 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 
     @Transactional
     @Override
-    public void remover(Cidade cidade) {
-        cidade = buscar(cidade.getId());
+    public void remover(Long cidadeId) {
+        Cidade cidade = buscar(cidadeId);
+        if(cidade == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cidade);
     }
 }
