+++ src/main/java/com/amauri/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.amauri.algafood.infrastructure.repository;
 
 import com.amauri.algafood.domain.model.Estado;
 import com.amauri.algafood.domain.repository.EstadoRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,8 +39,11 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
     @Transactional
     @Override
-    public void remover(Estado estado) {
-        estado = buscar(estado.getId());
+    public void remover(Long id) {
+        Estado estado = buscar(id);
+        if(estado == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(estado);
     }
 }
